style: added user role controll before destroy only for admin

+++ app/Http/Controllers/Admin/AllergenController.php
@@ -103,7 +103,11 @@ class AllergenController extends Controller
      */
     public function destroy(Allergen $allergen)
     {
-        //
+
+        if (auth()->user()->role !== 'admin') {
+            abort(403);
+        }
+
         $allergen->delete();
 
         return redirect()->route('allergens.index');
